Move resourc-y things to 'api', 'ResourceError' to 'connector'.
