+++ api/src/main/scala/quasar/api/datasource/DatasourceError.scala
@@ -17,7 +17,7 @@
 package quasar.api.datasource
 
 import slamdata.Predef._
-import quasar.common.resource.ResourcePath
+import quasar.api.resource.ResourcePath
 
 import monocle.Prism
 import scalaz.{Cord, Equal, ISet, NonEmptyList, Show}
