+++ impl/src/main/scala/quasar/impl/evaluate/Source.scala
@@ -16,7 +16,7 @@
 
 package quasar.impl.evaluate
 
-import quasar.common.resource.ResourcePath
+import quasar.api.resource.ResourcePath
 
 import monocle.macros.Lenses
 import scalaz.{Applicative, Cord, Equal, Order, Show, Traverse}
