+++ repl/src/main/scala/quasar/repl/ReplState.scala
@@ -20,7 +20,7 @@ package repl
 import slamdata.Predef._
 
 import quasar.api.datasource.DatasourceType
-import quasar.common.resource.ResourcePath
+import quasar.api.resource.ResourcePath
 import quasar.fp.numeric.widenPositive
 
 import eu.timepit.refined.api.Refined
