+++ run/src/main/scala/quasar/run/QuasarError.scala
@@ -18,8 +18,8 @@ package quasar.run
 
 import slamdata.Predef.{Exception, Product, Serializable, Throwable}
 import quasar.api.datasource.DatasourceError.CreateError
-import quasar.common.resource.ResourceError
 import quasar.compile.SemanticErrors
+import quasar.connector.ResourceError
 import quasar.qscript.PlannerError
 import quasar.sql.ParsingError
 import quasar.yggdrasil.vfs.{ResourceError => MimirResourceError}
