+++ run/src/main/scala/quasar/run/implicits.scala
@@ -17,9 +17,9 @@
 package quasar.run
 
 import quasar.api.datasource.DatasourceError.CreateError
-import quasar.contrib.scalaz.MonadError_
-import quasar.common.resource.ResourceError
 import quasar.compile.SemanticErrors
+import quasar.connector.ResourceError
+import quasar.contrib.scalaz.MonadError_
 import quasar.qscript.PlannerError
 import quasar.yggdrasil.vfs.{ResourceError => MimirResourceError}
